Add script of Francisco Torreira
The praatfig repository now includes two  scripts which are located into two different subfolders. The first one 'Beginner' is for very biginner to learn how to use a simple script for produce a figure on Praat. The second one 'DrawObjects_byFranciscoTorreira' is an advance script from Francisco Torreira to draw objects in separate frames.